Added entity loading from mods

+++ UnityPoC/Assets/Scripts/GameObjects/EntityInfo.cs
@@ -9,7 +9,7 @@ public struct Entity {
 
         public string id;
 
-        public double speed;
+        public float speed;
 
         public string spriteName;
